Documentation: Fix the trailing whitespaces #5618
It is a good practice to remove trailing whitespaces from the
files. They can not be spottet in most editors, and they add no
value. This commit removes all trailing whitespaces from the repository.

The pre-commit hook `trailing-whitespaces` takes care of trailing
white-spaces before the commit. This is a desired behavior.