Documentation: Fix the file endings #5618
It is a good practice to remove empty lines from the files. They can not
be spottet in most editors, and they add no value. This commit removes
all unnecessary empty lines on the end of files from the repository.

The pre-commit hook `end-of-file-fixer` takes care of empty lines on the
end of files before the commit. This is a desired behavior.

There, however, is the `ISSUETEMPLATE.md` file, where the behavior is
desired. We do not change that one.