Documentation: Improve the error message for the missing type annotations Actions job #4455
The missing type annotations job in the GitHub Actions fails, if there
are more missing type annotations after the push/PR than before. This
ensures, that new code is type annotated.

While the "HOW" is answered (online recourses, other developers, common
sense, ...), the "WHY" is not. New developers or the ones not familiar
with our desire to add type annotations to the code-base might ask
themselfes why we are doing this and if that is a mistake. We should
therefore improve the error message and explain, why we are doing this.

This commit links the Rucio Type Annotations Guide proposal to the error
message thrown by a failing missing type annotations job. This way,
developers unfamiliar with the topic, can research why we are doing
this.