Core & Internals: migrate to pymemcache. Closes #5341
Dogpile only started supporting pymemcache in 1.1.2, so increase
the pinned version.

Compared to the python-memcached backend, pymemcache doesn't silently
ignore errors when the backend is not available. As a result, with
pymemcache, rucio raises exceptions when run in an environment
where memcached is not available. Detect this case and explicitly
disable the caching layer.