Core & Internals: explicitly set limit(1) on existence queries in did.py
the behavior of first() (as consequence, scalar() too) was changed
between the session.query() and session.execute() constructs.
In the sqlalchemy2 case, all rows are fetched, and the first row is
selected on python side:
https://docs.sqlalchemy.org/en/14/core/connections.html#sqlalchemy.engine.Result.first

This results in a huge performance regression due to our usage of
first() as an existence check.