Monitoring & Logging: add datatype to transfer events. #4557
The commit takes quite a dirty approach of retrieving the `datatype`
directly via a sql query when preparing the message. There is
no easy way to improve that:
- In all code path, there is a lot of logic between the moment
when we retrieve the work queue from the database and the moment
when we sent the message. Forwarding the datatype through all the
call stack will make the code more complicated.
- We cannot import from core.dids here, because it creates a circular
import problem. So using existing get_metadata calls is not easily
achievable to avoid a raw database call.