Monitoring & Logging: fix transfer_link sent by receiver. Closes #5640
Under normal operation, when we get a notification from FTS about a
transfer completion, we know about the transfer and retrieve it
from the database. It doesn't cost us anything to use the
request.external_host field from the db table, which has the correct
format.

We can ignore the occurances when it's not the case. It means that we
got a notification from FTS about a transfer submitted by rucio, but
about which rucio doesn't know anything.... we would have bigger
problems to investigate in such case than a wrong monitoring link.