Tests: support mocking message table
It allows to more easily verify the content of messages generated
during a test.

Re-use the same mechanism which was used for mocking the config table.
Move the common code to a generic function allowing to create any
in-memory table and use that code in both places.

Had to re-work the registration in the sessionmaker: it is now
possible to have both a Message and a Config table mock used in
parallel and this was not supported by previous code.