Core & Internals: reduce verbosity of cache access errors
Otherwise we print the stack trace all the time when we run the
rucio client.

Also, check if it's a client running and completely disable the cache
in those cases. is_client() is not 100% accurate, as it will still
return False when the rucio client is run from a machine configured
to run a server/daemon. For example, in the dev containers the
logging messages will still be printed, but should be less scary now.