Core & Internals: rework temp table usage
This commit mostly reworks the pipeline for the oracle database, but
also touches a bit the postgresql logic.

On oracle, private temporary tables now have a random suffix appended,
to allow migration of core/dids to usage of temporary tables.
Actually, did core functions are sometimes called multiple times in the
same transaction. For example: multiple attach_dids calls. This results
in an attempt to create multiple times the same table. Oracle doesn't
support the "create table if not exists" construct, and retrieving
the list of private temporary tables is not natively supported by
sqlalchemy, so this is the best solution I had in mind. This shouldn't
result in an ifinite grow of number of tables, because these tables are
automatically dropped on commit.

Still on oracle, we now support having global temporary tables. The code
will automatically detect the existence of such tables and switch to
using them over dynamically created private temporary tables. Allow to
use a separate, more generic, name for the global ones.

For the same reason of multiple attach_dids being called in the same
session, set if_not_exists=True when creating the postgresql tables.

On databases which support it, we now add indexes to temporary tables.
Some delete and update queries which rely on exists() checks in the
temporary table should, in theory, benefit from it, but other
operations may suffes an aditional overhead for maintaining the indexes.