Core & Internals: Add Availability translation class #5664
The availability is stored as a bit-sequence in an integer. The three
least significant bits represent the respective read, write and delete
availabilities. While this saves space (compared to three full byte
bools plus metadata) and is quick to calculate, it is difficult to read
and error prone. The `Availability` class adds an easy to use interface
to make up the converison between the integer and the corresponding
values.

The `__iter__` method allows us to unpack the class. This corresponds to
the tuple unpacking and provides an easy way of accesssing the
individual values.

Since `Availability` is a class, the access for individual values is
more descriptive than a tuple (e.g. `availability[1]` vs
`availability.write`).

The `Availability` class can not be marked as a dataclass, since
`dataclasses` are added to Python in 3.7 and we have to support Python
3.6.

The default value of `Availability` should be None. This represents the
missing value. If a RSE is created, this should evaluate to `True`. If a
RSE is updated, None entries should not change the existing value.