Transfers: Fix FTS3 `checksum_validation_strategy`; Closes #5670
The FTS3 transfertool wrapper incorrectly determined which checksumming
algorithms to apply based on RSE attributes. In one case where the
source RSE supported checksums, but the destination RSE did not, the
determined strategy would be to perform no checksumming at all, instead
of selecting the source as checksum target.

This PR simplifies the logic in this function considerably and adds test
to verfiy behaviour.