Transfers: fix checksum validation in fts
Merge fts-specific function utils/set_checksum_value into the fts3.py
This way, we directly generate the checksum in the format accepted by
FTS in the calling context.

Fix the parse_checksum_support_attribute function. It was assuming
that attributes is a list, while it is a string, meaning that this
code just couldn't work. Moreover, it didn't handle correctly empty
attributes and would return [''].
Also, remove an unused related function in rse.py.

FTS doesn't support 'verify_checksum' = 'destination'. The correct
term to use is "target". Fix this.

Remove 'verify_checksum' from the fts "file" dictionary. This field
is ignored. Only the one from job_params is important. The same for
'request_type', remove it from the fts "file" dictionary (but keep
it in the file_metadata dictionary).

Add an e-2-e test to ensure fts correctly understands the parameters
that we set.