Transfers: get rid of dict_attributes in the Transfer object
This dict was a leftover from times when we used dictionaries to pass
data through the transfer stack. Now that the checksum handling was
cleaned up, the only dirty hack still relying on this old behavior is
'selection_strategy'. With a small code move we can get rid of this
hacky behavior.