Cache: Add the ignore_arguments_key_generator function #5692
Dogpile, our used cache abstraction, offers the feature to use the
`CacheRegion.cache_on_arguments` decorator to cache function results
based on a key created from all positional arguments of the
function. However, dogpile does not accept keywords in the function
signatures of the wrapped funtions.

We primairly use the cache to cache database lookups. These functions
need a database session, which we pass in as an additional _keyword_
argument. Since we don't want to include this session in the cache key,
the `ignore_arguments_key_generator` function is created. It is passed
to `CacheRegion.cache_on_arguments` as the `function_key_generator`
argument, and allows us to ignore function keyword arguments in the key
generation.