Feature 5591 create rules for suspicious replicas with only one declaration (#5653)
* Recovery: Create_rules_for_suspicious_replicas_with_only_one_declaration

It isn't possible to handle suspicious replicas that only have one trace, since there is too little information to go off of. Therefore, a rule for each such replica will be created, which will create multiple traces. This will cause the replica to be picked up by the suspicious replica recoverer, if the replica (or the RSE it is on) does indeed have a problem.

* Add Python annotations