Replicas: use temp tables for list_replicas
Rework queries in list_replicas to rely on temporary tables and
recursive CTE on database side for resolving dids.

The dids are then joined with the replica table and the filters
(rse_expression, ignore availability) applied when joining with
the replica table. Use a subquery to allow outerjoin against it
and get rid of the "files_wo_replica" logic which seems to be a
very hacky way to say that some file doesn't have replicas. Now
it's more directly returned by the query, but this has to be
evaluated performance wise.