Revert "Core & Internals: Split RSE availability column #5664"
This reverts commit 86bee483f50e7044cad44d4254f52458a909ec47.

The change in the database (alembic and models.py) is not reverted, to
maintain backward compatibility.

This commit introduces another migration strategy for the RSE
availability, since the old one could override some data. This is the
case if an older rucio version updates the `availablilty` column, while
the new one only look into the new columns. A migration in the end could
delete some data.

The new strategy has the following steps:

1. Add the new columns to the database and write the data to it, don't
   read from it. This creates the new columns, and fills them with
   intact data.

2. After all instances use the new version, use the new columns.

3. After all instances use the new column, delete the old one.