Core: Resolve the config not found exception #5704
Importing the client without a config file raises a RuntimeError:

```python
>>> from rucio.client import Client
...
RuntimeError: Could not load Rucio configuration file. Rucio looked in
the following paths for a configuration file, in order:
        /opt/rucio/etc/rucio.cfg
```

This should not happen, since the client is configurable _after_ the
import, and we thus do not need the configuration file for the import.