Database: Set the default availability_* value in the migration Vol 2. #5664
The default value for the Oracle db is `NONE`, even for the boolean
columns. This should be true for every column, to represent the default value 7
from the availability column.

Just setting the models default value is not enough, the column
definition also needs to be updated. Thank you Oracle for noticing,
thanks to myself for being so incompetent.