Monitoring & Logging: Ensure consistent units of time; Fix #5688
This PR aims to make the timer monitoring consistent with regards to
recording elapsed seconds vs. milliseconds (#5688). A new `Stopwatch`
class is introduced which handles the actual timing and recording in the
correct units, and the `record_timer_block` class is re-implemented using
the new `Stopwatch`.

All uses of `record_timer` and `record_timer_block` are then checked and
refactored if needed to ensure consistent units.