Database: Refactor the query result to dict functionality #5722
Sqlalchemy query results integrate some functionalities to communicate
changes to the database (e.g. update fields when they are updated in the
code). While this is desired and helpful, it is not always needed. If a
dict is passed to an external dependency or returned as write only, the
communication with the database is not needed.

Converting a query result to a dict is not trivial. Calling the `dict`
function on the result returns not just the desired fields, but also
some custom ones from `sqlalchemy` (for now `_sa_instance_state`, maybe
more in the future). Remove the `sqlalchemy` related fields and add a
test to make sure only these are returned.

This commit refactors the dict from sqlalchemy result creation into its
own function. This centralizes the functionality and decreases the
possibility for errors.