Protocols: Migrate WebDAV XML parsing to Python's ElementTree; Fix #5700, fix #5308
As described in #5700, the parser for the WebDAV server responses had
issues with namespace handling. This caused Rucio to misinterpret
responses from for instance the Apache DAV module. Changing the
implementation to use Python's xml.etree.ElementTree fixes this issue.

This PR also includes a fix for #5308, in which Rucio doesn't properly
handle DAV servers with path prefixes.