Core & Internals: reorganize deletion order in delete_dids
To minimize the number of insertions and deletions in the temp table,
group queries depending on what dids they work on: all/file/collection.
If delete_dids is called only on one type of dids(either files or
collections), the temporary table will only be filled once and the data
in it re-used throughout all the delete_dids execution.