Replicas: resolve file dids into a temp table
Instead of resolving the datasets into a temporary table, followed
by a query to resolve files from these datasets, directly resolve
the files. This is done entirely on the database side and kept
in a temporary table.

Also, optimize out un-needed parts of list_replica queries:
before doing any actual list replica calls, perform a fast check
of the input to decide if it's actually needed to perform all the
complex queries.