Replicas: opportunistically apply nrandom on database side. Closes #5418
list_replicas performs some filtering in python code after the
replicas are fetched from the database. For this reason, it's not
possible to reliably select n random replicas via a database query.
Some selected replicas may be dropped afterwards, leaving us with
less than n replicas to return to the client.

However, this commit tries to optimize the most common case of using
nrandom: downloading a handful of files from a very big collection.
In such case it is worth trying to pre-filter and fetch a small subset
of rows. If python-side filtering leaves us without the needed
number of replicas, just fall-back to fetching everything.