Core: Fix rse update availability can't be set to all false #5744
The `availability` attribute of an rse should be variable. `update_rse`
is the core function to do that. While updating the `availability` to a
non-zero value (at least one of `availability_{read,write,delete}` is
`True`) works, updating it to zero does not.

This comes due to an invalid check in line 1434, introduced in 8b033b393
(Core & Internals: Throw an exception on wrong RSE update option #5491,
2022-05-03). The variable `availability` is set to zero, which is a
valid entry in this case. The check in line 1434 however tries to use
the `availability` property of `param`, which does not exist.  This
check was introduced to also account for the direct update of the
`availability` from a number, not through the individual booleans. This
is needed for the `import_rses` in the importer.

This commit changes the priority to the following and thus gets rid of
the exception:

1. The availability number stored in the availability field

2. The individual availability values

3. The default value

We can not guarantee that only one of `availability` and
`availability_{read,write,delete}` should be updated since the recent
introduction of the `availability_{read,write,delete}` columns in the
RSE model. The importer takes all of the columns and tries to update
them.