Replicas: rework domain usage in list_replicas
Avoid duplicate code when constructing schemes and protocols by
iterating over a previously constructed list of desired domains.

Get rid of deepcopy: `domain` is a string, so assignment is enough
in this context. We don't do any in-place modification of that string.

Use the domain of the protocol when constructing the pfn: this is a
behavioral change introduced by this commit, which seems more correct.
In case when list_replicas is called with domain = 'all', we convert
it into two separate domains: 'wan' and 'lan'. Each of these 2 domains
can have its own protocol. Until this change, because the input domain
'all' was used in the 'if' check, the 'wan' pfn would be generated
without a proxy. With this change, the 'wan' pfn originated from the
domain = 'all' will have the server-side root proxy configured
(when relevant).