Replicas: move list_replicas' rse_info retrieval closer to usage
rse_info is only needed the first time we encounter a rse in list_rses.
It allows to construct a list of protocols.
There is thus no need to keep a dictionary with all rse_infos: it's not
needed later in the code.

This also allows to get rid of a very hacky behavior where rse_info
was used to store temporary variables only needed for protocol creation:
scheme priorities for lan and wan. Just use a local variable for that.