Tools: Pyright static type checking; Fix #5727
This commit introduces the Pyright type checker to the CI pipeline. The
CI check works in a similar way to the current "flake8-annotations"
check, in that it compares the number of typing errors before and after
applying an incoming pull request. The check fails if the number of
errors has increased. The added value of this check is to check for
typing errors, instead of just counting missing annotations as in the
"flake8-annotations" check. These two checks are orthogonal to each
other and serve different purposes. Pyright to a certain degree does
type inference, but produces better results in presence of type
annotations. The annotations themselves serve little value if there is
no type checker to use them except for increased readability. Therefore,
the "flake8-annotations" check will continue to be active.