Monitoring & Logging: implement rucio log formatter. #5573
The formatter should be a drop-in replacement to the python builtin
formatter, with two additional additions:
- it can output directly to json
- it can include data from the flask 'request' object into the format

When the logger writes to a json format, it tries to respect the
Elastic Common Schema (ECS) specification, but without getting too
strict about it.

When the format string contains a dot-separated "path" starting with
`http.request.`, the rucio formatter will try to extract the given
path from the flask `request` object.

Both these changes should allow to simplify the log ingestion pipeline.
- Json simplifies injection, removing the need to have complex parsing
of log lines and tokenization + merging of multiple lines into one log.
- Request data, like the client ip, can now be collected from within
the wsgi application instead of parsing httpd logs to extract it.