Tools: run_pyright.sh: comparisons with previous commits; Fix #5758
This commit adds the subcommands "generate-for-commit" and
"compare-with-commit" to "tools/run_pyright.sh". These actions act
similar to "generate" and "compare", except that they operate on a
specific git commit instead of the current work tree. Specifically,
"generate-for-commit" checks out the given commit before generating the
Pyright report as usual. "compare-with-commit" generates two reports,
one for the given commit and one for the current work tree, then
compares these as usual, like "compare" would.

To prevent modifications to the main git worktree, these new actions
will generate a temporary linked git worktree checked out at the given
commit, which is removed after the operation. This requires git version
> 2.18.0.