Rules: Provide more feedback on StatementErrors #5766
It is falsely assumed that the `StatementError` in `update_rule` is
always coming from a badly formatted rule id. It is an error that
occurred during execution of a SQL statement.

We pass more information to the caller to provide a better context of
the error. This makes the debugging for non-statement related errors
easier.