Transfers: gracefully skip missing sources in poller/receiver
The only difference is that the "except" block will not be triggered
and the stack-trace will not be included in the log message.

On the couple of cases I investigated, the missing sources were due to
the poller working on a request which is handled by receiver+finisher
in the meantime.