Tests: expose random_dataset_did and random_container_did in did_factory
This enables test factories to create a scope/name in correct format
without registering it to the database, but taking care of the
cleanup if ever the test function registers it.

This is mostly useful for non-atlas tests where these kind of dids
have different naming conventions. So being able to quickly generate
a did in correct naming helps.