Tests: don't use pre-defined MOCK rses in test_rule.py
For most tests, nothing to be done. For many it's just a simple
substitution of the rse name. The only exception is the
global account limit test which requires cleanup of global
account limits after the test to avoid impacting other tests.
For that, make use of the existing cleanup code by moving it
from test_account_limits.py to conftest.py.