Transfers: introduce core/topology
Mostly code move. Create a core/topology module which will be in
charge of RSE graph traversal and path computation.
Also rework the _RseLoaderContext usage. Move it to core/rse and
make it more generic. Don't bind it to a session anymore and be
always explicit about loading rse_data rather than relying on an
implicit operation.