Transfers: compute transfer paths in preparer too
Preparer takes very strong decisions about transfers:
 - throttle or not
 - which transfertool to use
while having limited knowledge about the topology and transfer.
This resulted, multiple times in bad decisions. For example, assigning
a transfer to the wrong transfertool.

This PR is a first step in doing the path computation and
decision-making in preparer instead of submitter.