Transfers: first try prepared source, then fallback to full path search
As preparer is much more intelligent now, the source it picks and
sets into request.source_rse_id has a high chance to be the one which
we want to use. Notable exceptions are:
- when the transfer took a long time to go from a preparing state to
  a queued state (for example: throttling). In that case we want to
  perform a full analysis of all available paths again, as the selected
  source may not be optimal anymore.
- when we detect that this path cannot actually be submitted. Submitter
  has more knowledge about transfertool specifics. For example, it can
  detect that a globus_endpoint_id attribute is not set on the rse.
  This also can happen in case of miss-configuration between preparer
  and submitter.

Change the submitter to only perform path computations from all sources
when source_rse_id is not set or the request was created longtime ago.

The PR reorganizes a bit the code by moving some constructs inside the
"Topology" class. For example: the list of multihop rses seems logical
to be part of the topology. This allows easier code re-use in both
submitter and preparer.