Transfers: some pyright and type hint fixes
In some cases, code has to be adapted because pyright wouldn't detect
correctly that the initialization of two variables is linked.
For example, if self.last_time is not None, self.logger also cannot
be None in hearbeat handler. So this doesn't fix any bugs, it's just
making things explicit for pyright to be happy.