Tests: use rse_factory in judge tests
Judges tests leave leftover requests in the databse.
Because of that, other tests fail. In particular,
the submitter tests which rely on the default submitter
"bulk" of 100.
Until now things "worked" because test_requests.py was
blast-removing all requests from the dabase. Effectively
"cleaning" up behind judge tests. But that behavior was
removed from test_requests.py in a previous commit.