Transfers: check transfertools for all hops in preparer
Until this commit, preparer was only checking that source rse and
destination rse had a common transfertool. This behavior is obviously
wrong. In case of multi-hop transfers, we can have the following path:
RSE1(fts3)->RSE2(fts3,globus)->RSE3(globus). The source and destination
RSEs don't have a common transfertool, but the transfer is possible.