CI: Disable the type checker for actions with no parent #5727
The type checker needs a base version to compare the current changes
to. Since the schedule actions does not provide such an attribute, we
should disable it.

The release branch is altered quite heavily from master, comparing it
agains the last mutual state results to many differences. Since
"in-between" commits which alter the types of different functions might
be missing we should disable it in general.