Transfers: fix multi-source in preparer->submitter interaction
Because submitter used only the source prepared by the preparer,
the multi-source submission was broken by the current preparer
and submitter rework. Fix it by ensuring that submitter considers
other sources for multi-source, but doesn't perform path computation
on them.

Disable multi-source search in preparer. We don't need that data for
now.

Don't compress multihop paths in preparer. This masks the real source
and works badly with source_replica expression. Because preparer
sets in the database a source_rse_id which doesn't match the expression.
The submitter was correctly recovering from this state, but this
resulted in more work than needed.