Transfers: rework throttler limit handling. Closes #5901
The commit does many changes to throttler internal workflow while
keeping the externally visible changes at minimum.

We get rid of the throttler/mode configuration parameter. It was used
to decide if throttler should work per activity or per all activities
+ if source throttling or destination throttling is desired. Now
throttler can work in all modes at the same time to respect whatever
limits are set in the database. Preparer also will automatically put
transfers in Waiting state if there are limits applying either to the
source or to the destination RSE.

Don't remove transfer limits anymore once all waiting requests are
handled. That behavior was counter-intuitive and undesired.

Limits now apply in a certain 'direction'. This didn't require a schema
change, because the transfer_limits table already contains the needed
field, but with a small restriction: 'direction' can only take one value
per primary key (rse_id/activity), so the same rse cannot apply both
source and destination throttling. This can easily be fixed with a
schema upgrade.

Throttler automatically finds out if there are any limits configured
on the RSEs involved in the transfer and for the concrete activity.
If more than one limit exists, all of them are enforced. For example:
- the source rse only allows 5 outgoing transfers for activity A
- the destination rse only allows 20 inbound transfers for any activity
will result into 5 transfers of activity A to be released, plus 15 more
from other activities.

This commit prepares the ground for limits which apply on multiple RSEs
by an RSE expression. For example: if a limit allows 1000 transfers from
'*', and RSE A already has 500 active transfers, another rse, B,
matching the same expression will only be allowed 500. To finalize the
support of this functionality, a database change and additional
development will be needed.