Tests: get rid of setup_class fixture in throttler tests
It is more explicit to have a fixture which return a connected
rse pairs rather than a fixture which runs at class initialization
and sets some class attributes.

Moreover, the setup_class doesn't work well with the pyright type
checker, who isn't able to understand that the class has the required
attributes and hence raises a warning.