Transfers: rework `transfertool` RSE attribute usage. Closes #5825
Until this commit, to allow a certain transfertool on an RSE, two
distinct attributes had to be set: 1) the `transfertool` attribute
which could be a list (ex: globus,fts3); and 2) the transfertool-
specific configuration attribute (ex: fts=<list_of_servers>,
globus_endpoint_id=<some_uuid>).
The 'transfertool' attribute is usually redundant. So it was cumbersome
to set it for each RSE, and it was decided that not setting it at all
means 'all transfertools are allowed'.
However, some queries required this attribute to be set if
the filter_transfertool configuration option was set in submitter.

To get rid of this confusing behavior, we decided to get rid of the
'transfertool' attribute usage in conveyor all together. Now, each
transfertool defines a list of 'required' rse attributes. If all those
attributes are set, it is considered that the corresponding transfertool
is enabled on the RSE. To enable FTS it's not enough to set the `fts`
attribute to a non-empty value. Similarly, globus requires
`globus_endpoint_id` to be set.