Core & Internals: lazy account_usage counter creation. Closes #5914
Until this commit, whenever an RSE or an account was created, the
account_usage table was pre-filled with empty counters to maintain
the cartesian product between the account and rse tables.
This means that most counters are empty. In atlas, only 0.5% of
them has a non-empty value.

This commit re-works the workflow to not fill the table with empty
counters and consider missing rows as a counter with value `0`.

The primary sponsor for this change are the unit tests, where we create
and teardown test accounts and RSEs quite frequently.