Transfers: rework throttler core/database interface. Closes #5805
This commit completely breaks the compatibility with limits previously
set in the database.

Limits are now in their separate table. And apply to an rse_expression.
All rses resolving to that expression are now subject to a common limit.
For performance reasons, the rse expression is converted to rows
in a new rse_transfer_limits table, which now is just a join-table
between `transfer_limits` and `rses` tables.

The core interface will now automatically fill this join-table whenever
a certain limit is set or updated. The re-evaluation of rse expressions
is also done by the throttler at the start of each run.