Monitoring & Logging: rework metric names. Closes #5804
All metrics are now prefixed with the module name in which they are
declared.

To achieve this goal, I initially envisioned automatic function
introspection (using the builtin module "inspect") and patching,
but I quickly realised that it will definitely be confusing to
have metric names which are "automatically" prepended with prefixes.

I settled on a very explicit and easy way to achieve this goal:
a module global variable which is initialized whenever the module is
imported. This requires some boilerplate code in each module file,
such as: 1) initializing this global object; 2) calling all metric
record function via this global object; but is very explicit and
should be much more difficult to miss-use.

Make all MultiMetric classes private. The only interface which must
be used now to get a metric is the MetricManager object which is in
charge of correctly prefixing all metrics. This interface now returns
MultiMetric instantiated objects, so we have to explicitly call the
corresponding labels/inc/set/observe methods of the metric to actually
record the value.

Get rid of the Timer() object and the "divisor" logic in it.
Use a stopwatches and explicit divisions instead. I consider that
it's not justified to have this complex abstraction to avoid
an explicit division in <10 places around the code base.