Common: Refactor execute to single return signature better typehints; Fix #5942
The function rucio/common/utils.py::execute is of great importance for
the testing workflows, since it allows for shell commands to be run and
their outputs to be used. The issue was that the function had two return
signatures, depending on the value of a keyword argument
(blocking=True). The different return signatures cause linting errors.

The function was refactored to have a single return signature, removing
the keyword argument and only treating the case where blocking=True. The
one situation in which the function 'execute' is called with
blocking=False has been handled by substituting the body of the
'execute' function into what was the function call.