Core & Internals: make session a mandatory keyword-only argument. #5947
The pyright type checker outputs 500-1000 errors of the format
"execute is not a member of None".

This behavior is correct. we explicitly mark the function as
accepting session=None, but then use it inside the function as if
it cannot be None. In reality, marking session as "Optional[Session]"
is actually completely wrong. The decorated function always expects
a valid session. However, the function returned by the decorator
has a different signature, it allows an empty session.

This commit makes Session mandatory for all functions decorated by
session decorators. To keep the risk of breaking code at minimum,
I avoid re-ordering function parameters. Hopefully, session is usually
the last argument and is almost always called as a keyword argument
in rucio. Enforce this behavior by making session (and logger)
a keyword-only argument. This way, it's possible to require a value
for this argument while keeping parameter reordering at minimum.

This change required to decorate did_meta functions with the session
decorators. Otherwise, calling those functions without session passed
raises an exception. It feels like those decorators where just
forgotten in the first place anyway.